+++ src/musikcore/library/MasterLibrary.h
@@ -56,7 +56,7 @@ namespace musik { namespace core { namespace library {
             virtual ~MasterLibrary();
 
             int Enqueue(QueryPtr query, Callback cb = Callback()) override;
-            int EnqueueAndWait(QueryPtr query, int64_t timeoutMs = kWaitIndefinite, Callback cb = Callback()) override;
+            int EnqueueAndWait(QueryPtr query, size_t timeoutMs = kWaitIndefinite, Callback cb = Callback()) override;
             musik::core::IIndexer *Indexer() override;
             int Id() override;
             const std::string& Name() override;
